Make all four software pages private. They cannot be accessed directly. Instead, they may be accesed requesting them via RequestPageIntoView
