+++ src/UniGetUI/MainWindow.xaml.cs
@@ -21,6 +21,7 @@ using Windows.ApplicationModel.DataTransfer;
 using Microsoft.Windows.AppNotifications;
 using UniGetUI.Core.Classes;
 using UniGetUI.Interface.Enums;
+using UniGetUI.Interface.SoftwarePages;
 using UniGetUI.PackageEngine.PackageClasses;
 using UniGetUI.Pages.DialogPages;
 
@@ -159,7 +160,9 @@ namespace UniGetUI.Interface
 
             if (action == NotificationArguments.UpdateAllPackages)
             {
-                NavigationPage.UpdatesPage.UpdateAll();
+                foreach (IPackage package in PEInterface.UpgradablePackagesLoader.Packages)
+                    if (package.Tag is not PackageTag.BeingProcessed and not PackageTag.OnQueue)
+                        MainApp.Operations.Update(package);
             }
             else if (action == NotificationArguments.ShowOnUpdatesTab)
             {
@@ -253,11 +256,11 @@ namespace UniGetUI.Interface
                 if (Id != "" && CombinedManagerName != "" && ManagerName == "" && SourceName == "")
                 {
                     Logger.Warn($"URI {link} follows old scheme");
-                    NavigationPage.DiscoverPage.ShowSharedPackage_ThreadSafe(Id, CombinedManagerName);
+                    DialogHelper.ShowSharedPackage_ThreadSafe(Id, CombinedManagerName);
                 }
                 else if (Id != "" && ManagerName != "" && SourceName != "")
                 {
-                    NavigationPage.DiscoverPage.ShowSharedPackage_ThreadSafe(Id, ManagerName, SourceName);
+                    DialogHelper.ShowSharedPackage_ThreadSafe(Id, ManagerName, SourceName);
                 }
                 else
                 {
@@ -330,8 +333,9 @@ namespace UniGetUI.Interface
                     {
                         // Handle potential JSON files
                         Logger.ImportantInfo("Begin attempt to open the package bundle " + param);
-                        NavigationPage.NavigateTo(PageType.Bundles);
-                        _ = NavigationPage.BundlesPage.OpenFromFile(param);
+
+                        var page = NavigationPage.RequestPageIntoView<PackageBundlesPage>(PageType.Bundles);
+                        _ = page.OpenFromFile(param);
                     }
                     else if (param.EndsWith("UniGetUI.exe") || param.EndsWith("UniGetUI.dll"))
                     {
