+++ src/UniGetUI/Pages/SoftwarePages/DiscoverSoftwarePage.cs
@@ -296,92 +296,5 @@ namespace UniGetUI.Interface.SoftwarePages
 
         private void MenuInstallSettings_Invoked(object sender, RoutedEventArgs e)
             => ShowInstallationOptionsForPackage(SelectedItem);
-
-        public void ShowSharedPackage_ThreadSafe(string id, string combinedSourceName)
-        {
-            var contents = combinedSourceName.Split(':');
-            string managerName = contents[0];
-            string sourceName = "";
-            if (contents.Length > 1) sourceName = contents[1];
-            ShowSharedPackage_ThreadSafe(id, managerName, sourceName);
-        }
-
-        public void ShowSharedPackage_ThreadSafe(string id, string managerName, string sourceName)
-        {
-            MainApp.Instance.MainWindow.DispatcherQueue.TryEnqueue(async () =>
-            {
-                IPackage? package = await GetPackageFromIdAndManager(id, managerName, sourceName);
-                if (package is not null) ShowDetailsForPackage(package);
-            });
-        }
-
-        private static async Task<IPackage?> GetPackageFromIdAndManager(string id, string managerName, string sourceName)
-        {
-            try
-            {
-                Logger.Info($"Showing shared package with pId={id} and pSource={managerName}: ´{sourceName} ...");
-                MainApp.Instance.MainWindow.Activate();
-                DialogHelper.ShowLoadingDialog(CoreTools.Translate("Please wait...", id));
-
-                IPackageManager? manager = null;
-
-                foreach (var candidate in PEInterface.Managers)
-                {
-                    if (candidate.Name == managerName || candidate.DisplayName == managerName)
-                    {
-                        manager = candidate;
-                        break;
-                    }
-                }
-
-                if (manager is null)
-                {
-                    throw new ArgumentException(CoreTools.Translate("The package manager \"{0}\" was not found", managerName));
-                }
-
-                if(!manager.IsEnabled())
-                    throw new ArgumentException(CoreTools.Translate("The package manager \"{0}\" is disabled", manager.DisplayName));
-
-                if(!manager.Status.Found)
-                    throw new ArgumentException(CoreTools.Translate("There is an error with the configuration of the package manager \"{0}\"", manager.DisplayName));
-
-                var results = await Task.Run(() => manager.FindPackages(id));
-                var candidates = results.Where(p => p.Id == id).ToArray();
-
-                if (candidates.Length == 0)
-                {
-                    throw new ArgumentException(CoreTools.Translate("The package \"{0}\" was not found on the package manager \"{1}\"", id, manager.DisplayName));
-                }
-
-                IPackage package = candidates[0];
-
-                // Get package from best source
-                if (candidates.Length >= 1 && manager.Capabilities.SupportsCustomSources)
-                    foreach (var candidate in candidates)
-                        if (candidate.Source.Name == sourceName)
-                            package = candidate;
-
-                Logger.ImportantInfo($"Found package {package.Id} on manager {package.Manager.Name}, showing it...");
-                DialogHelper.HideLoadingDialog();
-                return package;
-            }
-            catch (Exception ex)
-            {
-                Logger.Error($"An error occurred while attempting to show the package with id {id}");
-                Logger.Error(ex);
-                var warningDialog = new ContentDialog
-                {
-                    Title = CoreTools.Translate("Package not found"),
-                    Content = CoreTools.Translate("An error occurred when attempting to show the package with Id {0}", id) + ":\n" + ex.Message,
-                    CloseButtonText = CoreTools.Translate("Ok"),
-                    DefaultButton = ContentDialogButton.Close,
-                    XamlRoot = MainApp.Instance.MainWindow.Content.XamlRoot // Ensure the dialog is shown in the correct context
-                };
-
-                DialogHelper.HideLoadingDialog();
-                await MainApp.Instance.MainWindow.ShowDialogAsync(warningDialog);
-                return null;
-            }
-        }
     }
 }
